+++ shapash/report/smart_report/report_styles.css
@@ -1,0 +1,285 @@
+:root {
+    --shapash-gold: #ffbb00;
+    --text-main: #333;
+    --text-light: #777;
+}
+
+*,
+*::before,
+*::after {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+}
+
+html {
+    scroll-behavior: smooth;
+}
+
+body,
+.report-shell {
+    background: #fdfdfd;
+    color: var(--text-main);
+    font-family: 'Helvetica Neue', Arial, sans-serif;
+}
+
+.report-shell {
+    display: flex;
+    min-height: 100vh;
+    width: 100%;
+}
+
+.sidebar {
+    width: 240px;
+    background: #fff;
+    border-right: 1px solid #eee;
+    position: fixed;
+    height: 100vh;
+    padding: 30px 20px;
+    overflow-y: auto;
+    box-shadow: inset -8px 0 12px -12px rgba(0, 0, 0, 0.45);
+}
+
+.sidebar-brand {
+    margin-bottom: 40px;
+    display: flex;
+    align-items: center;
+    gap: 10px;
+    color: var(--shapash-gold);
+    font-size: 18px;
+    font-weight: bold;
+}
+
+.sidebar-brand-logo {
+    display: block;
+    width: 34px;
+    height: 34px;
+    flex: 0 0 auto;
+}
+
+.sidebar-brand-text {
+    line-height: 1.2;
+}
+
+.nav-item {
+    color: var(--text-light);
+    padding: 10px 0;
+    display: block;
+    text-decoration: none;
+    font-size: 13px;
+    transition: 0.2s;
+}
+
+.nav-item:hover {
+    color: var(--text-main);
+}
+
+.nav-item.active {
+    color: #551a8b;
+    font-weight: bold;
+    border-left: 3px solid #551a8b;
+    padding-left: 10px;
+}
+
+.nav-group {
+    margin-bottom: 2px;
+}
+
+.nav-group-title {
+    color: var(--text-main);
+    font-size: 13px;
+    font-weight: 500;
+    padding: 10px 0;
+    display: block;
+    text-decoration: none;
+    transition: 0.2s;
+}
+
+.nav-group-title:hover {
+    color: #551a8b;
+}
+
+.nav-group-title.active {
+    color: #551a8b;
+    font-weight: bold;
+    border-left: 3px solid #551a8b;
+    padding-left: 10px;
+}
+
+.nav-children {
+    display: block;
+}
+
+.nav-child {
+    color: var(--text-light);
+    padding: 6px 0 6px 16px;
+    display: block;
+    text-decoration: none;
+    font-size: 12px;
+    transition: 0.2s;
+}
+
+.nav-child:hover {
+    color: var(--text-main);
+}
+
+.nav-child.active {
+    color: #551a8b;
+    font-weight: 600;
+    border-left: 3px solid #551a8b;
+    padding-left: 13px;
+}
+
+.container {
+    margin-left: 240px;
+    width: 100%;
+    padding: 60px 80px 60vh;
+    max-width: 1200px;
+}
+
+.main-header {
+    text-align: center;
+    margin-bottom: 60px;
+}
+
+.main-header h1 {
+    font-size: 2.4rem;
+    font-weight: 500;
+    color: #000;
+    margin-bottom: 20px;
+}
+
+.section-title {
+    font-size: 1.6rem;
+    color: #000;
+    margin: 40px 0 20px;
+    font-weight: 700;
+}
+
+.section-block {
+    margin-bottom: 30px;
+}
+
+.content-block {
+    margin-bottom: 30px;
+    line-height: 1.6;
+    font-size: 14px;
+}
+
+.shapash-callout {
+    border-left: 4px solid var(--shapash-gold);
+    background: #fff;
+    padding: 15px 25px;
+    margin: 30px 0;
+    color: #333;
+    line-height: 1.6;
+    font-size: 15px;
+}
+
+.kv-table,
+table.dataframe {
+    width: 100%;
+    border-collapse: separate;
+    border-spacing: 0;
+    margin: 12px 0 24px;
+    background: #fff;
+    border: 1px solid #ececec;
+    border-radius: 12px;
+    overflow: hidden;
+    box-shadow: 0 8px 24px rgba(0, 0, 0, 0.04);
+}
+
+.kv-table thead th,
+table.dataframe thead th {
+    background: #fafafa;
+    color: #000;
+    font-weight: 700;
+    text-align: left;
+    border-bottom: 1px solid #ececec;
+    padding: 12px 16px;
+}
+
+.kv-table tbody th,
+table.dataframe tbody th {
+    color: #000;
+    font-weight: 600;
+    text-align: left;
+    background: #fcfcfc;
+}
+
+.kv-table td,
+.kv-table th,
+table.dataframe td,
+table.dataframe th {
+    padding: 12px 16px;
+    border-bottom: 1px solid #f1f1f1;
+    vertical-align: top;
+}
+
+.kv-table tbody tr:last-child td,
+.kv-table tbody tr:last-child th,
+table.dataframe tbody tr:last-child td,
+table.dataframe tbody tr:last-child th {
+    border-bottom: 0;
+}
+
+.kv-table tbody tr:nth-child(even) td,
+table.dataframe tbody tr:nth-child(even) td {
+    background: #fdfdfd;
+}
+
+.kv-key {
+    font-weight: 700;
+    width: 220px;
+    color: #000;
+    white-space: nowrap;
+}
+
+.kv-key-label,
+.kv-key-sep {
+    white-space: nowrap;
+}
+
+.kv-val {
+    color: var(--text-main);
+}
+
+.badge {
+    display: inline-block;
+    padding: 6px 14px;
+    border: 1px solid #eee;
+    border-radius: 4px;
+    font-size: 12px;
+    background: #fff;
+}
+
+.shapash-divider {
+    border-bottom: 1px solid #eee;
+    margin: 50px 0;
+}
+
+.scroll-section {
+    scroll-margin-top: 40px;
+}
+
+.panel-plot {
+    width: 100%;
+    overflow-x: auto;
+}
+
+.panel-plot .bk-root,
+.panel-plot .plotly-graph-div {
+    width: 100%;
+    max-width: 100%;
+}
+
+@media (max-width: 900px) {
+    .sidebar {
+        display: none;
+    }
+
+    .container {
+        margin-left: 0;
+        padding: 30px 40px 40vh;
+    }
+}
